requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.5
openai==1.82.1
SQLAlchemy==2.0.41
python-dotenv==1.1.0
PyPDF2==3.0.1
pytest==7.4.0
httpx==0.27.0
